pyproject.toml
[project]
name = "aill25"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.8",
    "gradio>=5.16.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pydantic-settings>=2.7.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]
